Habemus Loveleditor files loaded in game !
Now the game can scan a LovelEditor generated file and add all the objects in game.
If the tiles are properly named and in the correct order in the file, it can also draw the tile's image.

The LevelEditor file must be altered to add the enemies and player's info (sounds,positions)!!