Fixed the art loading part
Now the game searches for the correct number (in the string inside levels/3/leveltest_fodase_3.lua) and use this number as an index to load the proper art.

The number and the order of "love.graphics.newImage" still matters. Example : plat_3.png must be in the third position of the respective table to be loaded properly.